+++ app/controllers/passkit/dashboard/passes_controller.rb
@@ -1,0 +1,9 @@
+module Passkit
+  module Dashboard
+    class PassesController < ApplicationController
+      def index
+        @passes = Passkit::Pass.order(created_at: :desc).includes(:devices).first(100)
+      end
+    end
+  end
+end