+++ test/dummy/config/initializers/passkit.rb
@@ -1,0 +1,2 @@
+Passkit.configure do |_config|
+end