+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## [0.4.0]
+
+- Allow to use the dashboard also in production.
+- Allow to protect the dashboard using different strategies. Basic auth is default.
+- Breaking: now your Passkit dashboard is mounted under `/passkit/dashboard` instead of just `/passkit`. 
+
 ## [0.3.3]
 
-- Fix previews page
+- Fix previews page.
 
 ## [0.3.2]
 
@@ -12,4 +18,4 @@
 
 ## [0.1.0]
 
-- Initial release
+- Initial release.